some crosstalk api work
git-svn-id: http://svn.psychdesktop.net/trunk@297 815afad1-63b8-44ff-8d19-411ecde00fa5

+++ backend/api/crosstalk.php
@@ -22,19 +22,22 @@
     {
     header('Content-type: text/xml');
 	include("../config.php");
-        $appID = $_GET['appID'];
+    $appID = $_GET['appID'];
+	$instance = $_GET['instance'];
 	$query = "SELECT * FROM ${db_prefix}crosstalk WHERE userid=\"${userid}\" AND appID=\"${appID}\"";
 	$link = mysql_connect($db_host, $db_username, $db_password) or die('Could not connect: ' . mysql_error());
 	mysql_select_db($db_name) or die('Could not select database');
 	$output = "<" . "?xml version='1.0' encoding='utf-8' ?" . ">\r\n" . "<crosstalkEvents>";
 	$result = mysql_query($query) or die('Query failed: ' . mysql_error());
 		while ($row = @mysql_fetch_array($result, MYSQL_ASSOC)) {
+		if($row["instance"] == $null || $row["instance"] == $instance || $row["instance"] == 0) {
 		$appid = $row["appID"];
 		$sender = $row["sender"];
 		$message = $row["message"];
-		$output .=  "\r\n" . '<event sender="'. $row["sender"] . '" appID="'. $row["appID"] .'">'. $row["message"] .'</event>';
+		$output .=  "\r\n" . '<event sender="'. $row["sender"] . '" appID="'. $row["appID"] .'" instance="'. $row["instance"] .'">'. $row["message"] .'</event>';
 		$query = "DELETE FROM ${db_prefix}crosstalk WHERE userid=\"${userid}\" AND appID=\"${appid}\" AND message=\"${message}\"";
 		mysql_query($query) or die('Query failed: ' . mysql_error());
+		}
 		}		
 	$output .=  "\r\n" . "</crosstalkEvents>";	
 	echo($output);
@@ -46,8 +49,17 @@
     $sender = $userid;
     $destination = $_GET["destination"];
     $appID = $_GET["appID"];
-    $query = "INSERT INTO `${db_prefix}crosstalk` (userid, message, sender, appID) VALUES('${sender}', '${message}', '${destination}', '${appID}');";
-    $link = mysql_connect($db_host, $db_username, $db_password) or die('Could not connect: ' . mysql_error());
+	$instance = $_GET["instance"];
+	if($destination == 0) {
+	$destination = $sender;
+	}
+	if($instance != $null) {
+	$query = "INSERT INTO `${db_prefix}crosstalk` (userid, message, sender, appID, instance) VALUES('${sender}', '${message}', '${destination}', '${appID}', '${instance}');";
+    }
+	else {
+	$query = "INSERT INTO `${db_prefix}crosstalk` (userid, message, sender, appID) VALUES('${sender}', '${message}', '${destination}', '${appID}');";
+    }
+	$link = mysql_connect($db_host, $db_username, $db_password) or die('Could not connect: ' . mysql_error());
     mysql_select_db($db_name) or die('Could not select database');
     mysql_query($query) or die('Query failed: ' . mysql_error());
     echo("OK.");
